python-api/requirements.txt
Flask
pymongo
python-dotenv
bson
Flask-CORS